+++ lib/concierge/suppliers/ciirus/entities/property_permissions.rb
@@ -3,7 +3,7 @@ module Ciirus
     class PropertyPermissions
       attr_reader :property_id, :mc_enable_property, :agent_enable_property, :agent_user_id, :mc_user_id,
                   :native_property, :calendar_sync_property, :aoa_property, :time_share,
-                  :online_booking_allowed
+                  :online_booking_allowed, :deleted
 
       def initialize(attrs = {})
         @property_id            = attrs[:property_id]
@@ -16,6 +16,7 @@ module Ciirus
         @aoa_property           = attrs[:aoa_property]
         @time_share             = attrs[:time_share]
         @online_booking_allowed = attrs[:online_booking_allowed]
+        @deleted                = attrs[:deleted]
       end
 
       def to_h
@@ -29,7 +30,8 @@ module Ciirus
           calendar_sync_property: calendar_sync_property,
           aoa_property:           aoa_property,
           time_share:             time_share,
-          online_booking_allowed: online_booking_allowed
+          online_booking_allowed: online_booking_allowed,
+          deleted:                deleted
         }
       end
     end