+++ lib/concierge/suppliers/ciirus/validators/permissions_validator.rb
@@ -8,6 +8,7 @@ module Ciirus
     #   * online booking is not allowed
     #   * property is timeshare (as well as GetReservations method doesn't support such properties)
     #   * properties with AOA(Allocation On Arrival) Bookings Mode
+    #   * property deleted
     #
     class PermissionsValidator
       attr_reader :permissions
@@ -17,7 +18,10 @@ module Ciirus
       end
 
       def valid?
-        online_booking_allowed? && !timeshare? && !allocation_on_arrival?
+        online_booking_allowed? &&
+          !timeshare? &&
+          !allocation_on_arrival? &&
+          !property_deleted?
       end
 
       private
@@ -32,6 +36,10 @@ module Ciirus
 
       def allocation_on_arrival?
         permissions.aoa_property
+      end
+
+      def property_deleted?
+        permissions.deleted
       end
     end
   end