+++ spec/lib/concierge/suppliers/ciirus/commands/property_permissions_fetcher_spec.rb
@@ -13,6 +13,7 @@ RSpec.describe Ciirus::Commands::PropertyPermissionsFetcher do
   let(:property_id) { {property_id: '33680'} }
 
   let(:success_response) { read_fixture('ciirus/responses/property_permissions_response.xml') }
+  let(:deleted_property_response) { read_fixture('ciirus/responses/deleted_property_permissions_response.xml') }
   let(:error_response) { read_fixture('ciirus/responses/error_property_permissions_response.xml') }
   let(:wsdl) { read_fixture('ciirus/wsdl.xml') }
 
@@ -56,6 +57,16 @@ RSpec.describe Ciirus::Commands::PropertyPermissionsFetcher do
         expect(permissions.aoa_property).to be_falsey
         expect(permissions.time_share).to be_falsey
         expect(permissions.online_booking_allowed).to be_truthy
+        expect(permissions.deleted).to be_falsey
+      end
+
+      it 'fills deleted property permissions with right deleted attribute' do
+        stub_call(method: described_class::OPERATION_NAME, response: deleted_property_response)
+
+        result = subject.call(property_id)
+
+        permissions = result.value
+        expect(permissions.deleted).to be_truthy
       end
     end