+++ CHANGELOG.md
@@ -11,6 +11,7 @@ how this file is formatted and how the process works.
 
 ### Fixed
 - Kigo property images with non-Latin characters
+- Ciirus:: skip deleted properties during sync
 
 ## [0.8.0] - 2016-09-08
 ### Added